requirements.txt
# Core dependencies - pinned versions
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
yfinance==0.2.43
pandas_market_calendars==4.4.2
pyyaml==6.0.2
matplotlib==3.8.4
scikit-learn==1.4.2
loguru==0.7.2
pyarrow==16.1.0

# Additional dependencies for data fetching
requests==2.31.0
lxml==5.0.0
html5lib==1.1
beautifulsoup4==4.12.0

# Visualization
seaborn==0.12.0

# Development and testing
pytest==8.1.1
pytest-mock==3.12.0

# Code quality tools
ruff==0.6.9
mypy==1.11.2

# Jupyter notebooks (optional, for analysis)
# jupyter==1.0.0
# ipykernel==6.25.0
# notebook==7.0.0

# Progress bars
tqdm==4.66.2
